+++ CHANGELOG.md
@@ -21,3 +21,12 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+## v1.0.0
+
+Released Friday 5th April 2024.
+
+### New
+
+- added `roundDown()` function
+- added `roundUp()` function
+- added `roundTo()` function